requirements.txt
Flask==3.0.0
Flask_RESTful==0.3.10
ipython==8.16.1
Pillow==10.0.1
Pillow==10.0.1
Requests==2.31.0
torch==2.1.0
tqdm==4.65.0
transformers==4.34.0
